Docint/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
docling>=2.5.0
openai
pdfminer.six
pymupdf